+++ src/pages/Home.tsx
@@ -2,16 +2,14 @@ import React from 'react';
 
 import GetPost from '../components/posts/GetPost';
 import Banner from '../components/common/Banner';
-import CategoryFilter from '../components/category/CategoryFilter';
-import InfiniteScroll from '../components/InfiniteScroll';
+// import InfiniteScroll from '../components/InfiniteScroll';
 
 const Home = () => {
   return (
     <div>
       <Banner />
-      <CategoryFilter />
       <GetPost />
-      <InfiniteScroll />
+      {/* <InfiniteScroll /> */}
     </div>
   );
 };
